requirements.txt
biopython
gff3
scipy
pybedtools
sklearn
bcbio-gff